requirements.txt
asgiref
Django
django-admin-interface
django-colorfield
django-modeladmin-reorder
flake8
flake8-broken-line
flake8-isort
flake8-plugin-utils
flake8-return
isort
mccabe
mysqlclient
pep8-naming
pillow
pybitrix24
pycodestyle
pyflakes
python-dotenv
python-slugify
sqlparse
text-unidecode
tzdata
